+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+
+- Added `loadDataContext` prop to `ChartProvider` for passing along state to the `loadData` function without creating new `Function` instances. [#37](https://github.com/palantir/react-layered-chart/issues/37).
+
 ## 1.5.1 (2016-10-11)
 
 ### Fixed